added uuid as primary key for todo

+++ src/main/java/com/nitin/todo/service/TodoHibernateService.java
@@ -2,11 +2,11 @@ package com.nitin.todo.service;
 
 import com.nitin.todo.dao.TodoHibernateDao;
 import com.nitin.todo.model.Todo;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 
 
 @Service
@@ -24,29 +24,38 @@ public class TodoHibernateService {
     }
 
     public List<Todo> getAllTodos() {
-        List<Todo> todoList = todoHibernateDao.findAll();
+
+        Iterator<Todo> todoItr = todoHibernateDao.findAll().iterator();
+        List<Todo> todoList = new ArrayList<>();
+        while (todoItr.hasNext())
+            todoList.add(todoItr.next());
+
         return todoList;
+
     }
 
-    public Optional<Todo> getTodoById(long id) {
-        Optional<Todo> todo = todoHibernateDao.findById(id);
-        return todo;
+    public Optional<Todo> getTodoById(UUID id) {
+        return todoHibernateDao.findById(id);
     }
 
-    public int updateTodoById(long id, Todo updatedTodo)throws Error {
-        Todo todoToUpdate = todoHibernateDao.findById(id).get();
-        if (todoToUpdate == null) {
-            throw new Error("Todo not found");
+    public int updateTodoById(UUID id, Todo updatedTodo) throws Exception {
+        if (todoHibernateDao.findById(id).isPresent()) {
+            Todo todoToUpdate = todoHibernateDao.findById(id).get();
+            todoToUpdate.setTitle(updatedTodo.getTitle());
+            todoToUpdate.setDescription(updatedTodo.getDescription());
+            todoHibernateDao.save(todoToUpdate);
+            return 0;
+        } else {
+            throw new Exception("Todo not found");
         }
-
-        todoToUpdate.setTitle(updatedTodo.getTitle());
-        todoToUpdate.setDescription(updatedTodo.getDescription());
-        todoHibernateDao.save(todoToUpdate);
-        return 0;
     }
 
-    public int deleteTodoById(long id){
-        todoHibernateDao.deleteById(id);
+    public int deleteTodoById(UUID id) throws NoSuchElementException {
+        if (todoHibernateDao.findById(id).isPresent()) {
+            todoHibernateDao.deleteById(id);
+        } else {
+            throw new NoSuchElementException("Todo not found");
+        }
         return 0;
     }
